move client to its own folder
