+++ src/DataAccess/IBigQueryClientFactory.cs
@@ -1,4 +1,6 @@
-﻿namespace sharpbq.DataAccess;
+﻿using sharpbq.DataAccess.Clients;
+
+namespace sharpbq.DataAccess;
 
 public interface IBigQueryClientFactory
 {
